fix(ucx): Align exchange unit tests with count-only flow and schema checks
- Update remainingBytes expectations for count-only fetches
  - What was broken:
    - v1RepeatedFetchAdvancesQueue, v2FetchesSequencesZeroAndOne, and
      v2OversizeRequestReturnsOneChunk still expected remainingBytes to
      report one pending chunk after a successful fetch.
    - Count-only UCX flow leaves remainingBytes empty and advances via
      later fetches or sequence numbers instead.
  - How it is fixed:
    - Assert remainingBytes is empty after the first fetch.
    - Keep advancement coverage through subsequent fetch and sequence
      checks already present in those cases.

- Align Narrow test tables with declared RowType
  - What was broken:
    - UcxExchangeTest Narrow helpers built cuDF columns as UINT32/FLOAT32
      while kTestRowType advertised INTEGER/DOUBLE (INT32/FLOAT64).
    - CudfVector::validatePhysicalSchema now rejects that mismatch, so
      the first Narrow basicTest aborted with:
      "CudfVector schema mismatch at column 0 (c0): Velox INTEGER ->
      cuDF 3, actual cuDF 7" and stopped the suite early (38/158).
  - How it is fixed:
    - Generate and verify Narrow columns as int32_t and double so host
      vectors, cuDF types, and RowType stay consistent.

- Verification:
  - Reproduced the aborting Narrow case, then confirmed it and the
    remaining Narrow basicTest plus UcxOutputQueueManagerTest /
    IntraNodeTransferRegistryTest suites pass after the fix.

+++ velox/experimental/ucx-exchange/tests/UcxTestData.cpp
@@ -263,8 +263,8 @@ void UcxTestData::initialize(
           << "]";
   numRows_ = numRows;
   strings_ = std::make_shared<std::vector<std::string>>();
-  integers_ = std::make_shared<std::vector<uint32_t>>();
-  floats_ = std::make_shared<std::vector<float>>();
+  integers_ = std::make_shared<std::vector<int32_t>>();
+  floats_ = std::make_shared<std::vector<double>>();
 
   std::random_device rd;
   std::mt19937 gen(rd());
@@ -293,10 +293,10 @@ std::unique_ptr<cudf::table> UcxTestData::makeTable(
     rmm::cuda_stream_view stream) {
   std::vector<std::unique_ptr<cudf::column>> columns;
 
-  // Column 0: INT32 (integers)
+  // Column 0: INT32 (integers) - matches kTestColumnTypes INTEGER()
   columns.push_back(makeNumericColumn(*integers_, stream));
 
-  // Column 1: FLOAT32
+  // Column 1: FLOAT64 - matches kTestColumnTypes DOUBLE()
   columns.push_back(makeNumericColumn(*floats_, stream));
 
   // Column 2: STRING
@@ -317,8 +317,8 @@ bool UcxTestData::verifyTable(
   }
 
   // Get the data from the received table
-  auto receivedInts = getColVector<uint32_t>(table.column(0), numRows, stream);
-  auto receivedDoubles = getColVector<float>(table.column(1), numRows, stream);
+  auto receivedInts = getColVector<int32_t>(table.column(0), numRows, stream);
+  auto receivedDoubles = getColVector<double>(table.column(1), numRows, stream);
   auto receivedStrings = getStringCol(table.column(2), numRows, stream);
 
   // Compare with expected data. Use modular indexing because batch
